requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.36.0
peft>=0.7.0
bitsandbytes>=0.41.0
accelerate>=0.25.0
datasets>=2.16.0

# OpenAI API
openai>=1.10.0
tiktoken>=0.5.2

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.66.0
loguru>=0.7.2

# Data processing
pandas>=2.0.0
numpy>=1.24.0
jsonlines>=4.0.0

# Evaluation
scikit-learn>=1.3.0
evaluate>=0.4.1

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: Monitoring and logging
wandb>=0.16.0
tensorboard>=2.15.0
